Abstract main window
