+++ Leader Key/Cheatsheet.swift
@@ -123,7 +123,7 @@ enum Cheatsheet {
         let screenHalf = screen.visibleFrame.width / 2
         let desiredWidth: CGFloat = 580
         let margin: CGFloat = 20
-        return min(desiredWidth, screenHalf - margin * 2 - mainViewSize / 2)
+        return min(desiredWidth, screenHalf - margin * 2 - mysteryBoxSize / 2)
       }
       return 580
     }
